From #GRIPMAR2021 TASK 1: Prediction Using Supervised ML
Objective:

Predict the percentage of a student based on the no. of study hours. What will be the predicted score if a student studies for 9.25 hrs/ day?

About the task: This task is Supervised Machine Learning(ML) and such kind of learning includes the training of the model on a labelled dataset that includes both training and validation of datasets. The labelled dataset is one that has both input and output parameters.